+++ plugins/org.eclipse.osee.orcs.db/src/org/eclipse/osee/orcs/db/internal/callable/CreateBranchDatabaseTxCallable.java
@@ -112,7 +112,7 @@ public class CreateBranchDatabaseTxCallable extends JdbcTransaction {
          int associatedArtifactId = newBranchData.getAssociatedArtifactId();
 
          // this checks to see if there are any branches that aren't either DELETED or REBASELINED with the same artifact ID
-         if (associatedArtifactId > -1 && associatedArtifactId != SystemUser.OseeSystem.getUuid()) {
+         if (associatedArtifactId > -1 && associatedArtifactId != SystemUser.OseeSystem.getUuid().intValue()) {
             int count = jdbcClient.runPreparedQueryFetchObject(connection, 0,
                "SELECT (1) FROM osee_branch WHERE associated_art_id = ? AND branch_state NOT IN (?, ?)",
                newBranchData.getAssociatedArtifactId(), BranchState.DELETED.getValue(),
